+++ lib/json_schemer/draft4/meta.rb
@@ -2,6 +2,12 @@
 module JSONSchemer
   module Draft4
     BASE_URI = URI('http://json-schema.org/draft-04/schema#')
+    FORMATS = Draft6::FORMATS.dup
+    FORMATS.delete('uri-reference')
+    FORMATS.delete('uri-template')
+    FORMATS.delete('json-pointer')
+    CONTENT_ENCODINGS = Draft6::CONTENT_ENCODINGS
+    CONTENT_MEDIA_TYPES = Draft6::CONTENT_MEDIA_TYPES
     SCHEMA = {
       'id' => 'http://json-schema.org/draft-04/schema#',
       '$schema' => 'http://json-schema.org/draft-04/schema#',