+++ lib/json_schemer/draft7/meta.rb
@@ -2,6 +2,11 @@
 module JSONSchemer
   module Draft7
     BASE_URI = URI('http://json-schema.org/draft-07/schema#')
+    FORMATS = Draft201909::FORMATS.dup
+    FORMATS.delete('duration')
+    FORMATS.delete('uuid')
+    CONTENT_ENCODINGS = Draft201909::CONTENT_ENCODINGS
+    CONTENT_MEDIA_TYPES = Draft201909::CONTENT_MEDIA_TYPES
     SCHEMA = {
       '$schema' => 'http://json-schema.org/draft-07/schema#',
       '$id' => 'http://json-schema.org/draft-07/schema#',