+++ lib/json_schemer/openapi30/meta.rb
@@ -2,6 +2,12 @@
 module JSONSchemer
   module OpenAPI30
     BASE_URI = URI('json-schemer://openapi30/schema')
+    # https://spec.openapis.org/oas/v3.0.3#data-types
+    FORMATS = OpenAPI31::FORMATS.merge(
+      'byte' => proc { |instance, _value| ContentEncoding::BASE64.call(instance).first },
+      'binary' => proc { |instance, _value| instance.is_a?(String) && instance.encoding == Encoding::ASCII_8BIT },
+      'date' => Format::DATE
+    )
     SCHEMA = {
       'id' => 'json-schemer://openapi30/schema',
       '$schema' => 'http://json-schema.org/draft-04/schema#',