+++ lib/json_schemer/openapi31/meta.rb
@@ -2,6 +2,14 @@
 module JSONSchemer
   module OpenAPI31
     BASE_URI = URI('https://spec.openapis.org/oas/3.1/dialect/base')
+    # https://spec.openapis.org/oas/v3.1.0#data-types
+    FORMATS = {
+      'int32' => proc { |instance, _format| instance.is_a?(Integer) && instance.bit_length <= 32 },
+      'int64' => proc { |instance, _format| instance.is_a?(Integer) && instance.bit_length <= 64 },
+      'float' => proc { |instance, _format| instance.is_a?(Float) },
+      'double' => proc { |instance, _format| instance.is_a?(Float) },
+      'password' => proc { |_instance, _format| true }
+    }
     SCHEMA = {
       '$id' => 'https://spec.openapis.org/oas/3.1/dialect/base',
       '$schema' => 'https://json-schema.org/draft/2020-12/schema',