+++ test/json_schema_test_suite_test.rb
@@ -45,7 +45,7 @@ class JSONSchemaTestSuiteTest < Minitest::Test
       path = Pathname.new(__dir__).join('..', 'JSON-Schema-Test-Suite', 'remotes', uri.path.gsub(/\A\//, ''))
       JSON.parse(path.read)
     else
-      JSON.parse(Net::HTTP.get(uri))
+      JSON.parse(fetch(uri))
     end
   end